requirements.txt
aniso8601==9.0.1
backcall==0.2.0
click==7.1.2
decorator==5.0.7
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
python-dotenv==0.17.0
pytz==2021.1
redis==3.5.3
six==1.15.0
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==1.0.1
